Amazon_scrapping_alert/requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
dnspython==2.1.0
idna==2.10
pymongo==3.11.2
python-dotenv==0.15.0
requests==2.25.1
selenium==3.141.0
soupsieve==2.1
urllib3==1.26.2
